support-files/lang-syncer/pyproject.toml
[project]
name = "lang-syncer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.1",
    "loguru>=0.7.3",
    "notion-client>=2.7.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]
